Get testinfra/app tests running against molecule
This was fun :| I removed some of the hacky yaml manipulation from the bash
scripts and instead am relying on environment variables. Also wanted to ensure
that the testinfra tests do not bomb out the whole job but continue to run so we
can get that sweet junit xml output.